+++ app/controllers/metadata_presenter/engine_controller.rb
@@ -113,35 +113,6 @@ module MetadataPresenter
     end
     helper_method :external_or_relative_link
 
-    def load_conditional_content
-      if @page.content_component_present?
-        if components_without_conditionals(@page).present?
-          return components_without_conditionals(@page) + show_components(@page).compact
-        end
-
-        show_components(@page).compact
-      end
-    end
-    helper_method :load_conditional_content
-
-    def components_without_conditionals(page)
-      page.content_components.select { |component|
-        component.conditionals.blank?
-      }.map(&:uuid)
-    end
-
-    def show_components(page)
-      return @page.content_components.map(&:uuid) if editor_preview?
-
-      page.content_components.map do |content_component|
-        EvaluateContentConditionals.new(
-          service:,
-          component: content_component,
-          user_data: load_user_data
-        ).show_component
-      end
-    end
-
     private
 
     def not_found
@@ -180,5 +151,18 @@ module MetadataPresenter
     def editor_preview?
       URI(request.original_url).path.split('/').include?('preview')
     end
+
+    def single_page_preview?
+      return false if in_runner?
+      return true if request.referrer.blank?
+
+      !URI(request.referrer).path.split('/').include?('preview')
+    end
+    helper_method :single_page_preview?
+
+    def in_runner?
+      ::Rails.application.class.module_parent.name == 'FbRunner'
+    end
+    helper_method :in_runner?
   end
 end