+++ spec/models/page_spec.rb
@@ -412,21 +412,27 @@ RSpec.describe MetadataPresenter::Page do
     end
   end
 
-  describe '#content_component_present?' do
-    context 'when page contains content components' do
-      let(:page) { service.find_page_by_url('check-answers') }
-
-      it 'should return true' do
-        expect(page.content_component_present?).to be_truthy
-      end
+  describe '#load_conditional_content' do
+    let(:service_metadata) { metadata_fixture(:conditional_content) }
+    let(:page) { service.find_page_by_url('content') }
+    let(:user_data) do
+      {
+        'multiple_radios_1' => 'Option A',
+        'multiple_checkboxes_1' => %w[2]
+      }
+    end
+    let(:expected_components) do
+      %w[
+        701f93e3-1d78-4a1f-9495-07e32b6e26fe
+        71bfc176-613d-42ec-8d53-63e50b696ec6
+        61139d00-53eb-4ff3-9227-6ecd0b80aac4
+        4d4d7ace-9ce2-4415-9edb-abcac75ed17b
+        6a3b4104-1d4c-4534-87d8-b81f5c764d43
+      ]
     end
 
-    context 'when page does not contain content components' do
-      let(:page) { service.find_page_by_url('holiday') }
-
-      it 'should return false' do
-        expect(page.content_component_present?).to be_falsey
-      end
+    it 'should return the right number of always display component' do
+      expect(page.load_conditional_content(service, user_data)).to eq(expected_components)
     end
   end
 end